Address TODO: renamed components
