+++ src/components/features/Settings/Settings.tsx
@@ -13,6 +13,7 @@ import DefaultLayout from "../../layouts/Default";
 import AppSettings from "./components/AppSettings";
 import ClickulatorSettings from "./components/ClickulatorSettings";
 import { FormattedMessage } from "react-intl";
+import Callout from "../../Callout";
 
 export const SettingsNavConfig: INavigationItem = {
   route: "/settings",
@@ -24,12 +25,9 @@ export default function Settings(props: RouteComponentProps) {
   return (
     <DefaultLayout>
       <form id="Settings" className="form">
-        <div className="b-callout">
-          <i className="material-icons b-callout__icon">info</i>
-          <p className="b-callout__blurb">
-            <FormattedMessage id="settings.info" />
-          </p>
-        </div>
+        <Callout>
+          <FormattedMessage id="settings.info" />
+        </Callout>
         <AppSettings />
         <ClickulatorSettings />
       </form>
